refactor: restructure hydration benchmark tests and configurations
