+++ .github/actions/benchmark/action.yml
@@ -5,10 +5,10 @@ runs:
     - name: Install dependencies
       uses: ./.github/actions/pnpm-setup
 
-    # - name: Install playwright dependencies and browsers
-    #   run: pnpm exec playwright install --with-deps chromium
-    #   shell: bash
-    #   working-directory: packages/tools/benchmark-tests
+    - name: Install playwright dependencies and browsers
+      run: pnpm exec playwright install --with-deps chromium
+      shell: bash
+      working-directory: packages/tools/benchmark-tests
 
     - name: Build required packages
       run: pnpm build:deps
