+++ .github/workflows/benchmark.pr-check.yml
@@ -21,11 +21,6 @@ jobs:
       - name: Install and execute benchmark
         uses: ./.github/actions/benchmark
 
-      - name: Load benchmark from target branch
-        run: |
-          git fetch origin ${{ github.base_ref }}
-          git show origin/${{ github.base_ref }}:packages/tools/benchmark-tests/benchmark-baseline.json > packages/tools/benchmark-tests/benchmark-baseline.json || echo "[]" > packages/tools/benchmark-tests/benchmark-baseline.json
-
       - name: Compare benchmark with baseline
         run: node scripts/compare-benchmark.mjs
         working-directory: packages/tools/benchmark-tests
