+++ packages/tools/benchmark-tests/.knip.json
@@ -1,15 +1,16 @@
 {
 	"$schema": "https://unpkg.com/knip@5/schema.json",
-	"entry": ["tests/*.spec.ts"],
-	"ignore": ["tests/benchmark.test.ts"],
+	"entry": ["**/*.ts"],
+	"ignore": ["public/{assets,components,theme}/**"],
 	"ignoreBinaries": ["build"],
 	"ignoreDependencies": [
 		"@public-ui/components",
 		"@public-ui/theme-default",
+		"@types/mocha",
+		"@wdio/globals",
 		"@wdio/local-runner",
 		"@wdio/mocha-framework",
 		"@wdio/spec-reporter",
-		"adopted-style-sheets",
 		"chromedriver",
 		"serve"
 	],
