+++ packages/tools/benchmark-tests/tests/lib/config.ts
@@ -1,0 +1,59 @@
+export const TEST_ITERATIONS = parseInt(process.env.TEST_ITERATIONS || '25', 10);
+export const TEST_TIMEOUT = parseInt(process.env.TEST_TIMEOUT || '5000', 10);
+// export const TEST_BATCH_SIZE = parseInt(process.env.TEST_BATCH_SIZE || '10', 10);
+export const TEST_URL = 'http://localhost:3000/test-page.html';
+
+export const TAGS = [
+	'kol-abbr',
+	'kol-accordion',
+	'kol-alert',
+	'kol-avatar',
+	'kol-badge',
+	'kol-breadcrumb',
+	'kol-button',
+	'kol-button-group',
+	'kol-button-link',
+	'kol-card',
+	'kol-details',
+	'kol-drawer',
+	'kol-form',
+	'kol-heading',
+	'kol-icon',
+	'kol-image',
+	'kol-indented-text',
+	'kol-input-checkbox',
+	'kol-input-color',
+	'kol-input-date',
+	'kol-input-email',
+	'kol-input-file',
+	'kol-input-number',
+	'kol-input-password',
+	'kol-input-radio',
+	'kol-input-text',
+	// 'kol-kolibri',
+	'kol-link',
+	'kol-link-button',
+	'kol-link-group',
+	'kol-logo',
+	'kol-modal',
+	'kol-nav',
+	'kol-pagination',
+	'kol-popover-button',
+	'kol-progress',
+	'kol-quote',
+	'kol-select',
+	'kol-skip-nav',
+	'kol-spin',
+	'kol-split-button',
+	// 'kol-symbol',
+	// 'kol-table',
+	'kol-table-stateful',
+	'kol-table-stateless',
+	'kol-tabs',
+	'kol-textarea',
+	// 'kol-toast-container',
+	'kol-toolbar',
+	'kol-tree',
+	'kol-tree-item',
+	'kol-version',
+] as const;
