+++ packages/tools/benchmark-tests/tests/lib/types.d.ts
@@ -1,0 +1,14 @@
+import { TAGS } from './config';
+
+export type TagType = (typeof TAGS)[number];
+
+export type Measure = {
+	hydrated: number | null;
+	themed: number | null;
+};
+
+export type Params = {
+	iterations: number;
+	tag: TagType;
+	timeout: number;
+};
