+++ packages/tools/benchmark-tests/tsconfig.json
@@ -1,0 +1,11 @@
+{
+	"compilerOptions": {
+		"target": "ES2020",
+		"module": "commonjs",
+		"lib": ["es2020", "dom"],
+		"strict": true,
+		"esModuleInterop": true,
+		"types": ["@wdio/globals/types", "mocha", "node"]
+	},
+	"include": ["./tests/**/*.ts"]
+}
